create classes for loads
